Moves polars code to script
