+++ python/mall/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ["MallFrame"]
 
-from mall.llm import MallFrame
+from mall.polars import MallFrame
